Attempt to upload photos from webcam to s3
This is a failed attempt at uploading photos to s3 with the
webcam. The main failure of this stems from the photo being coded
into base64. Previous attempts were made at getting the photo in
a better format.

This patch may be reintroduced later. But for now it remains as a
reference.